Pull v2 stuff from lib.rs to v2.rs
