+++ bindings_ffi/src/lib.rs
@@ -10,12 +10,6 @@ use logger::FfiLogger;
 use std::error::Error;
 use std::sync::{Arc, Mutex};
 use tokio::sync::{oneshot, oneshot::Sender};
-use xmtp_proto::api_client::{
-    BatchQueryResponse, PagingInfo, PublishResponse, QueryResponse, XmtpApiClient,
-};
-use xmtp_proto::xmtp::message_api::v1::IndexCursor;
-
-use xmtp_api_grpc::grpc_api_helper::Client as GrpcClient;
 
 use xmtp_api_grpc::grpc_api_helper::Client as TonicApiClient;
 use xmtp_mls::groups::MlsGroup;
@@ -25,10 +19,6 @@ use xmtp_mls::{
     builder::ClientBuilder,
     client::Client as MlsClient,
     storage::{EncryptedMessageStore, EncryptionKey, StorageOption},
-};
-use xmtp_proto::xmtp::message_api::v1::{
-    cursor::Cursor as InnerCursor, BatchQueryRequest, Cursor, Envelope, PublishRequest,
-    QueryRequest, SortDirection,
 };
 
 use crate::inbox_owner::RustInboxOwner;
@@ -63,20 +53,6 @@ fn stringify_error_chain<T: Error>(error: &T) -> String {
     }
 
     result
-}
-
-#[uniffi::export(async_runtime = "tokio")]
-pub async fn create_v2_client(
-    host: String,
-    is_secure: bool,
-) -> Result<Arc<FfiV2Client>, GenericError> {
-    let client = GrpcClient::create(host, is_secure).await?;
-
-    let client = FfiV2Client {
-        inner_client: Arc::new(client),
-    };
-
-    Ok(client.into())
 }
 
 #[uniffi::export(async_runtime = "tokio")]
@@ -158,308 +134,6 @@ impl FfiXmtpClient {
         self.inner_client.register_identity().await?;
 
         Ok(())
-    }
-}
-
-#[derive(uniffi::Enum)]
-pub enum FfiSortDirection {
-    Unspecified = 0,
-    Ascending = 1,
-    Descending = 2,
-}
-
-impl FfiSortDirection {
-    pub fn from_i32(val: i32) -> Self {
-        match val {
-            0 => FfiSortDirection::Unspecified,
-            1 => FfiSortDirection::Ascending,
-            2 => FfiSortDirection::Descending,
-            _ => panic!("Invalid sort direction"),
-        }
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiPagingInfo {
-    limit: u32,
-    cursor: Option<FfiCursor>,
-    direction: FfiSortDirection,
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiCursor {
-    pub digest: Vec<u8>,
-    pub sender_time_ns: u64,
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiV2QueryRequest {
-    content_topics: Vec<String>,
-    start_time_ns: u64,
-    end_time_ns: u64,
-    paging_info: Option<FfiPagingInfo>,
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiEnvelope {
-    content_topic: String,
-    timestamp_ns: u64,
-    message: Vec<u8>,
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiPublishRequest {
-    envelopes: Vec<FfiEnvelope>,
-    // ... the rest of the fields go here
-}
-
-impl From<FfiEnvelope> for Envelope {
-    fn from(env: FfiEnvelope) -> Self {
-        Self {
-            content_topic: env.content_topic,
-            timestamp_ns: env.timestamp_ns,
-            message: env.message,
-        }
-    }
-}
-
-impl From<Envelope> for FfiEnvelope {
-    fn from(env: Envelope) -> Self {
-        Self {
-            content_topic: env.content_topic,
-            timestamp_ns: env.timestamp_ns,
-            message: env.message,
-        }
-    }
-}
-
-impl From<PublishRequest> for FfiPublishRequest {
-    fn from(req: PublishRequest) -> Self {
-        Self {
-            envelopes: req.envelopes.into_iter().map(|env| env.into()).collect(),
-        }
-    }
-}
-
-impl From<FfiPublishRequest> for PublishRequest {
-    fn from(req: FfiPublishRequest) -> Self {
-        Self {
-            envelopes: req.envelopes.into_iter().map(|env| env.into()).collect(),
-        }
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiV2PublishResponse {}
-
-impl From<FfiV2PublishResponse> for xmtp_proto::xmtp::message_api::v1::PublishResponse {
-    fn from(_resp: FfiV2PublishResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<PublishResponse> for FfiV2PublishResponse {
-    fn from(_resp: PublishResponse) -> Self {
-        Self {}
-    }
-}
-
-impl From<SortDirection> for FfiSortDirection {
-    fn from(dir: SortDirection) -> Self {
-        match dir {
-            SortDirection::Unspecified => FfiSortDirection::Unspecified,
-            SortDirection::Ascending => FfiSortDirection::Ascending,
-            SortDirection::Descending => FfiSortDirection::Descending,
-        }
-    }
-}
-
-impl From<FfiSortDirection> for SortDirection {
-    fn from(dir: FfiSortDirection) -> Self {
-        match dir {
-            FfiSortDirection::Unspecified => SortDirection::Unspecified,
-            FfiSortDirection::Ascending => SortDirection::Ascending,
-            FfiSortDirection::Descending => SortDirection::Descending,
-        }
-    }
-}
-
-impl From<FfiCursor> for Cursor {
-    fn from(cursor: FfiCursor) -> Self {
-        Self {
-            cursor: Some(InnerCursor::Index(IndexCursor {
-                digest: cursor.digest,
-                sender_time_ns: cursor.sender_time_ns,
-            })),
-        }
-    }
-}
-
-fn proto_cursor_to_ffi(cursor: Option<Cursor>) -> Option<FfiCursor> {
-    match cursor {
-        Some(proto_cursor) => match proto_cursor.cursor {
-            Some(InnerCursor::Index(index_cursor)) => Some(FfiCursor {
-                digest: index_cursor.digest,
-                sender_time_ns: index_cursor.sender_time_ns,
-            }),
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-impl From<FfiV2QueryRequest> for QueryRequest {
-    fn from(req: FfiV2QueryRequest) -> Self {
-        Self {
-            content_topics: req.content_topics,
-            start_time_ns: req.start_time_ns,
-            end_time_ns: req.end_time_ns,
-            paging_info: req.paging_info.map(|paging_info| {
-                return PagingInfo {
-                    limit: paging_info.limit,
-                    direction: paging_info.direction as i32,
-                    cursor: paging_info.cursor.map(|c| c.into()), // TODO: fix me
-                };
-            }),
-        }
-    }
-}
-
-impl From<QueryRequest> for FfiV2QueryRequest {
-    fn from(req: QueryRequest) -> Self {
-        Self {
-            content_topics: req.content_topics,
-            start_time_ns: req.start_time_ns,
-            end_time_ns: req.end_time_ns,
-            paging_info: req.paging_info.map(|paging_info| {
-                return FfiPagingInfo {
-                    limit: paging_info.limit,
-                    direction: FfiSortDirection::from_i32(paging_info.direction),
-                    cursor: proto_cursor_to_ffi(paging_info.cursor),
-                };
-            }),
-        }
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiV2QueryResponse {
-    envelopes: Vec<FfiEnvelope>,
-    paging_info: Option<FfiPagingInfo>,
-}
-
-impl From<QueryResponse> for FfiV2QueryResponse {
-    fn from(resp: QueryResponse) -> Self {
-        Self {
-            envelopes: resp.envelopes.into_iter().map(|env| env.into()).collect(),
-            paging_info: resp.paging_info.map(|paging_info| {
-                return FfiPagingInfo {
-                    limit: paging_info.limit,
-                    direction: FfiSortDirection::from_i32(paging_info.direction),
-                    cursor: None,
-                };
-            }),
-        }
-    }
-}
-
-impl From<FfiV2QueryResponse> for QueryResponse {
-    fn from(resp: FfiV2QueryResponse) -> Self {
-        Self {
-            envelopes: resp.envelopes.into_iter().map(|env| env.into()).collect(),
-            paging_info: resp.paging_info.map(|paging_info| {
-                return PagingInfo {
-                    limit: paging_info.limit,
-                    direction: paging_info.direction as i32,
-                    cursor: None, // TODO: fix me
-                };
-            }),
-        }
-    }
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiV2BatchQueryRequest {
-    requests: Vec<FfiV2QueryRequest>,
-}
-
-#[derive(uniffi::Record)]
-pub struct FfiV2BatchQueryResponse {
-    responses: Vec<FfiV2QueryResponse>,
-}
-
-impl From<BatchQueryRequest> for FfiV2BatchQueryRequest {
-    fn from(req: BatchQueryRequest) -> Self {
-        Self {
-            requests: req.requests.into_iter().map(|req| req.into()).collect(),
-        }
-    }
-}
-
-impl From<FfiV2BatchQueryRequest> for BatchQueryRequest {
-    fn from(req: FfiV2BatchQueryRequest) -> Self {
-        Self {
-            requests: req.requests.into_iter().map(|req| req.into()).collect(),
-        }
-    }
-}
-
-impl From<BatchQueryResponse> for FfiV2BatchQueryResponse {
-    fn from(resp: BatchQueryResponse) -> Self {
-        Self {
-            responses: resp.responses.into_iter().map(|resp| resp.into()).collect(),
-        }
-    }
-}
-
-impl From<FfiV2BatchQueryResponse> for BatchQueryResponse {
-    fn from(resp: FfiV2BatchQueryResponse) -> Self {
-        Self {
-            responses: resp.responses.into_iter().map(|resp| resp.into()).collect(),
-        }
-    }
-}
-
-#[derive(uniffi::Object)]
-pub struct FfiV2Client {
-    inner_client: Arc<xmtp_api_grpc::grpc_api_helper::Client>,
-}
-
-#[uniffi::export(async_runtime = "tokio")]
-impl FfiV2Client {
-    pub async fn batch_query(
-        &self,
-        req: FfiV2BatchQueryRequest,
-    ) -> Result<FfiV2BatchQueryResponse, GenericError> {
-        let actual_req: BatchQueryRequest = req.into();
-        let result = self.inner_client.batch_query(actual_req).await?;
-        Ok(result.into())
-    }
-
-    pub fn set_app_version(&self, _version: String) {
-        log::info!("Needs implementation")
-    }
-
-    pub async fn publish(
-        &self,
-        request: FfiPublishRequest,
-        auth_token: String,
-    ) -> Result<FfiV2PublishResponse, GenericError> {
-        let actual_publish_request: PublishRequest = request.into();
-        let result = self
-            .inner_client
-            .publish(auth_token, actual_publish_request)
-            .await?;
-
-        Ok(result.into())
-    }
-
-    pub async fn query(
-        &self,
-        request: FfiV2QueryRequest,
-    ) -> Result<FfiV2QueryResponse, GenericError> {
-        let result = self.inner_client.query(request.into()).await?;
-        Ok(result.into())
     }
 }
